requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argparse-addons==0.12.0
bitstruct==8.20.0
cantools==40.2.1
click==8.3.0
contourpy==1.3.3
crccheck==1.3.0
cycler==0.12.1
diskcache==5.6.3
fastapi==0.117.1
fonttools==4.60.0
h11==0.16.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
msgpack==1.1.0
numpy==2.3.3
packaging==24.2
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-can==4.5.0
python-dateutil==2.9.0.post0
RPi.GPIO==0.7.1
six==1.17.0
sniffio==1.3.1
spidev==3.7
starlette==0.48.0
textparser==0.24.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uvicorn==0.36.0
wrapt==1.17.2
